docking_test/requirements.txt
# Core dependencies for molecular docking and visualization
biopython>=1.79
numpy>=1.21.0
matplotlib>=3.5.0
pandas>=1.3.0
plotly>=5.0.0
py3Dmol>=2.0.0

# Web interface dependencies
flask>=2.0.0
flask-cors>=3.0.0

# Optional dependencies (install via conda)
# pymol-open-source  # Install via: conda install -c conda-forge pymol-open-source
# rdkit              # Install via: conda install -c conda-forge rdkit
# openbabel          # Install via: conda install -c conda-forge openbabel